trasnmitting and listening states moved to managers:registries
