+++ src/rebroadcast/managers.py
@@ -1,3 +1,4 @@
+from collections import Counter
 from datetime import datetime, timedelta
 
 TIMEOUT = 5
@@ -5,7 +6,7 @@ TIMEOUT = 5
 class InUseFreq(Exception):
     pass
 
-class TransmitingState(object):
+class TransmissionRegistry(object):
 
     def __init__(self):
         self.active_freqs = dict()
@@ -29,5 +30,23 @@ class TransmitingState(object):
             del self.active_freqs[freq]
         except KeyError:
             pass
+
+class InternationalRegistry(object):
+
+    def __init__(self):
+        self.listening_freqs = Counter()
+
+    def add(self, freq):
+        
+        self.listening_freqs.update([freq])
+        
+        return self.listening_freqs[freq] == 1
+    
+    def remove(self, freq):
+        
+        if self.listening_freqs[freq] > 0:
+            self.listening_freqs.subtract([freq])
+        
+        return self.listening_freqs[freq] == 0
